requirements.txt
dj-database-url==0.4.1
Django==1.10.2
django-filter==0.15.3
djangorestframework==3.5.1
gunicorn==19.6.0
Markdown==2.6.7
Pillow==3.4.2
psycopg2==2.6.2
whitenoise==3.2
